requirements.txt
attrs==22.1.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
Bottleneck @ file:///C:/Windows/Temp/abs_3198ca53-903d-42fd-87b4-03e6d03a8381yfwsuve8/croots/recipe/bottleneck_1657175565403/work
certifi @ file:///C:/Windows/TEMP/abs_e9b7158a-aa56-4a5b-87b6-c00d295b01fanefpc8_o/croots/recipe/certifi_1655968940823/work/certifi
click==8.1.3
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1632508026186/work
colorama @ file:///C:/Windows/TEMP/abs_9439aeb1-0254-449a-96f7-33ab5eb17fc8apleb4yn/croots/recipe/colorama_1657009099097/work
contextlib2==21.6.0
debugpy @ file:///C:/ci/debugpy_1637091911212/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
execnet==1.9.0
Flask==2.2.2
importlib-metadata==5.0.0
iniconfig==1.1.1
ipykernel @ file:///C:/b/abs_21ykzkm7y_/croots/recipe/ipykernel_1662361803478/work
ipython @ file:///C:/Windows/TEMP/abs_45b5zb1l7q/croots/recipe/ipython_1659529855872/work
itsdangerous==2.1.2
jedi @ file:///C:/ci/jedi_1644297241925/work
Jinja2==3.1.2
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1616770841739/work
jupyter-core @ file:///C:/ci/jupyter_core_1651656282279/work
MarkupSafe==2.1.1
matplotlib-inline @ file:///C:/ci/matplotlib-inline_1661934089728/work
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186163140/work
mkl-service==2.4.0
mock==4.0.3
nest-asyncio @ file:///C:/ci/nest-asyncio_1649848126026/work
numexpr @ file:///C:/Windows/Temp/abs_e2036a32-9fe9-47f3-a04c-dbb1c232ba4b334exiur/croots/recipe/numexpr_1656940304835/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1653574840943/work
opencv-python==4.6.0.66
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas @ file:///C:/ci/pandas_1641461312795/work
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
path==16.5.0
path.py==12.5.0
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
pluggy==1.0.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
psutil @ file:///C:/Windows/Temp/abs_b2c2fd7f-9fd5-4756-95ea-8aed74d0039flsd9qufz/croots/recipe/psutil_1656431277748/work
py==1.11.0
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyqtgraph @ file:///tmp/build/80754af9/pyqtgraph_1594064840332/work
pyrealsense2==2.51.1.4348
pyserial==3.5
pytest==7.1.3
pytest-shutil==1.7.0
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///C:/Windows/TEMP/abs_90eacd4e-8eff-491e-b26e-f707eba2cbe1ujvbhqz1/croots/recipe/pytz_1654762631027/work
pywin32==302
pyzmq @ file:///C:/ci/pyzmq_1657597757589/work
sip @ file:///C:/ci/sip_1651028973764/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
spyder-kernels @ file:///C:/ci/spyder-kernels_1634237058681/work
termcolor==2.0.1
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tomli==2.0.1
tornado @ file:///C:/ci/tornado_1662476933490/work
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
typing_extensions==4.3.0
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
Werkzeug==2.2.2
wincertstore==0.2
zipp==3.8.1
